identitygc: Use int types for prometheus metrics
Even if (prometheus.Gauge).Set requires a float64 value, declaring the
counters for successful and unsuccessful gc runs as such might feel
confusing to the casual reader.  The type is then changed to int and a
type conversion to float64 is added at each Set call site.

+++ operator/identitygc/crd_gc.go
@@ -133,10 +133,10 @@ func (igc *GC) gc(ctx context.Context) error {
 	if igc.enableMetrics {
 		if ctx.Err() == nil {
 			igc.successfulRuns++
-			metrics.IdentityGCRuns.WithLabelValues(metrics.LabelValueOutcomeSuccess).Set(igc.successfulRuns)
+			metrics.IdentityGCRuns.WithLabelValues(metrics.LabelValueOutcomeSuccess).Set(float64(igc.successfulRuns))
 		} else {
 			igc.failedRuns++
-			metrics.IdentityGCRuns.WithLabelValues(metrics.LabelValueOutcomeFail).Set(igc.failedRuns)
+			metrics.IdentityGCRuns.WithLabelValues(metrics.LabelValueOutcomeFail).Set(float64(igc.failedRuns))
 		}
 		aliveEntries := totalEntries - deletedEntries
 		metrics.IdentityGCSize.WithLabelValues(metrics.LabelValueOutcomeAlive).Set(float64(aliveEntries))
